% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HHFindAge.R
\name{findagewcf}
\alias{findage}
\alias{findagewcf}
\title{Final HHMM18 Project}
\usage{
findage(academicyear, type = "average", department = "None",
  gender = "None", yearterminal = "None",
  TimebwBAandTerminalbyDept = "None", sumstats = FALSE,
  facultynumber = FALSE, stddev = FALSE)
}
\arguments{
\item{academicyear}{This function allows tour to input academic year in order to find the average age and other factors for that year}
}
\description{
Work Please
}

