\name{HHFindAge-package}
\alias{HHFindAge-package}
\alias{HHFindAge}
\docType{package}
\title{
Matt Morris '18 HH Final Project
}
\description{
Estimates age of williams college faculty
}
\details{
\tabular{ll}{
Package: \tab HHFindAge\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-01\cr
License: \tab GPL-2 \cr
}
Use them wisely
}
\author{
Matthew Morris
Maintainer: Matthew Morris <mrm10@williams.edu>

}
\references{

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
