% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacultyNumber.R
\name{wcagepredsfacultynumber}
\alias{facultynumber}
\alias{wcagepredsfacultynumber}
\title{facultynumber}
\usage{
facultynumber(academicyear, facultynumber = TRUE)
}
\arguments{
\item{academicyear}{This function allows a user to input academic year in order to find the average age and other characteristics for that year}

\item{facultynumber}{When set to True displays the number of faculty in the inputted academic year and when set to FALSE outputs the average age for that given academic year}
}
\description{
Allows a user to find the number of faculty at Williams College in a given academic year
}

