% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FindAge.R
\name{wcagepredsfindage}
\alias{findage}
\alias{wcagepredsfindage}
\title{findage}
\usage{
findage(academicyear, type = "average", department = "None",
  gender = "None", yearterminal = "None",
  TimebwBAandTerminalbyDept = "None")
}
\arguments{
\item{academicyear}{This function allows a user to input academic year in order to find the average age and other characteristics for that year}

\item{type}{This function allows a user to find average age, minimum age, or maximum age}

\item{department}{This function allows a user to define for a specific department what the average age is}

\item{gender}{This function allows a user to define for a given gender what the average age of the faculty is}

\item{yearterminal}{This function allows a user to find average age of the faculty by year of terminal degree}

\item{TimebwBAandTerminalbyDept}{This function allows a user to find the average time by Department it took for the faculty to receive their terminal degrees}
}
\description{
For a given academic year finds average age by department, youngest and oldest professor ages, average age by year of terminal degree, and years in between Bachelor's and Terminal Degree.
}

