% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stddevage.R
\name{wcagepredsstddevage}
\alias{stddevage}
\alias{wcagepredsstddevage}
\title{stddevage}
\usage{
stddevage(academicyear, stddevage = TRUE)
}
\arguments{
\item{academicyear}{This function allows a user to input academic year in order to find the average age and other characteristics for that year}

\item{stddevage}{When set to True displays the standard deviation of the Williams College faculty in the inputted academic year and when set to FALSE outputs the average age for that given academic year}
}
\description{
Allows a user to find the standard deviation of the faculty age distribution at Williams College in a given academic year
}

