% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SumStats.R
\name{wcagepredssumstats}
\alias{sumstats}
\alias{wcagepredssumstats}
\title{sumstats}
\usage{
sumstats(academicyear, sumstats = TRUE)
}
\arguments{
\item{academicyear}{This function allows a user to input academic year in order to find the average age and other characteristics for that year}

\item{sumstats}{When set to True displays the summary statistics of the Williams College faculty in the inputted academic year and when set to FALSE outputs the average age for that given academic year}
}
\description{
Allows a user to find the summary statistics such as the minimum, first quartile, median, mean, third quartile, and maximum of the faculty age distribution at Williams College in a given academic year
}

